requirements.txt
requests>=2.31.0
python-nmap>=0.7.1
pyyaml>=6.0
colorama>=0.4.6
